test: fix check_resname test
This test was relying on a "bad" alignment between two essentially
unrelated chains. This is no longer happening after the replacement of
GlobalAlign by SemiGlobal in commit ddfa2ab5, and the chains no longer
align meaningfully. As a result no residue names were mismatching, as
the representation of the binding site in the model no longer existed.

We removed this assumption and only replace a single residue in the
entity. This avoids depending on an alignment in this test.