chain mapping: modest refactor
Introduce GetRMSDMapping function which replaces GetRigidMapping

This is consistent with the other mapping functions that explicitely
contain score name in the function name: GetQSScoreMapping/GetlDDTMapping

Any GDT functionality has been dropped in the process. It never took off
and scoring is problematic anyways. Let's assume there is a huge complex
and you already mapped most of it. Now there is this chain which upon
superposition is 10A away and this other chain that is 1000A away.
GDT does not resolve this. They're both 0.0.