branch link refactor
Make sure that the mmcif info object has all branch link info, also
the ones connecting  atoms that are not resolved in the structure.
In general, this commit avoids that there is explicit reference to the
underlying structure in the info object, i.e. no actual AtomHandle objects
are stored in there just the raw info from the mmcif file.

Building the connectivity is delegated to MMCifReader.Parse