chain mapping: Avoid error for empty chem groups
Can occur in chain_mapping._ChainMappings that provides an iterator
of all possible ways to map mdl chains onto fixed ref chains.

The situation where a chem group (group of chains in the reference) is
empty but the respective group of chains in the model is not can be
correctly resolved.