enable pairwise sequence alignments with parasail
Parasail: https://github.com/jeffdaily/parasail

It's faster than the pairwise sequence algorithms we have in
OpenStructure. However, the main reason to check out another solution
is the fact that OpenStructure may return suboptimal alignment
results. This is a result of how the dynamic programming table is built.
If for a certain cell in the dynamic programming table a match state and
insertion state score equally, match is preferred. There is a chance
that we would get a better score later on if we would stick with the
insertion state (gap open penalties vs. gap extension penalties).
Not sure if this will be fixed in the near future.