Return None in case of no contacts in QS-score
The score is simply not valid and returning 0.0 as done before is incorrect.
Delegates further treatment of the problem to the caller which might want to
say: if both structures are monomer, I still give it a value of 1. Or, to
extend that definition, if I reproduced the correct stoichiometry, I still
give it a value of 1.