SCHWED-6298: QS-score for two monomers is defined as 1.0
They have the exact same quaternary structure in the end... This definition
has been generalized to: if there are no contacts observed, QS-score is 1.0
if the scored structures have the same stoichiometry, 0.0 otherwise.
This becomes a bit more abstract for higher order complexes but in principle
they still have the exact same quaternary structure if they match in
stoichiometry but have no single contact.