chain mapping: Use lDDT for mappings involving nucleotides - introduces optimized bb lDDT
QS-score which is used as default target function for chain mappings.
However, thats protein specific. As soon as nucleotides are involved,
lDDT with an increased inclusion radius of 30A is now used
in ChainMapper.GetMapping.

It was a bit embarassing that lDDT mappings were about an order of
magnitude slower than QS-score mappings. A specialized backbone only
lDDT has therefore been introduced that uses matrix operations.
This is not a replacement of the lDDTScorer. But it doesnt need to
produce per-residue scores and doesn't need to deal with symmetries
etc.