GDT: improve optimization and handle edge cases with low number of positions
Optimization has been improved by sampling several window sizes. The results
come really close to LGA (avg diff of 0.16 GDT points when evaluation all
CASP15 TS models).

In the previous implementation, an error was thrown when number of positions
was lower than the window size parameter. Now, the window size is just set
to the number of positions in that case and the algorithm runs through.
Special cases of only one or two positions are handled separately to produce
sensible output.