dockq: make results more similar to DockQ v2.2.1
lrmsd defines the longer chain as receptor and uses it for superposition.
If both chains have the same length, the selection becomes random.
This commit implements the same random selection as DockQ v2.2.1.
This is no conceptual change of the score itself! It just makes the
two implementations more similar.