scoring: force peptide/nucleotide connectivity based on residue numbers
If residue number alignments are enabled, one can assume that two
consecutive residues in terms of residue numbers are connected.
The conop Processor does not necessarily connect them if the bond
is considered unfeasible. This gives inaccurate results in
subsequent stereochemistry checks.
This change forces these connections if and only if resnum_alignments
are enabled