chain mapping: change definition of ChainMapper.chem_group_alignments
Old definition: List of alignments, one alignment for each chem group.
Each sequence in such an alignment was bound to an ATOMSEQ in the
target structure. The sequences were sorted by length and the first
one was considered the reference sequence.

New defition: List of alignments, one alignment for each chem group.
The first sequence is considered the reference sequence and is not
bound to any ATOMSEQ in the target structure. All subsequent sequences
are bound to an ATOMSEQ in the target structure. So if you derive these
alignments from structure only, the first sequence is simply a copy of
the longest ATOMSEQ sequence, essentially increasing the size of the
alignment by 1.

Motivation of this change is the ability of setting an arbitrary
reference sequence which can be the SEQRES.