scoring/chain mapping: rename unmapped_mdl_chains to mdl_chains_without_chem_mapping
While I agree that mdl_chains_without_chem_mapping is the longest
variable name ever, unmapped_mdl_chains suggests that is was not
mapped in the actual mapping procedure. However, with this variable
we're reporting the chains that didnt even make it into the
mapping procedure due to unsuccessful chem mapping.