ligand scoring: allow dist based connectivity heuristic for ligands from mmCIF
The default behaviour is to only allow ligands from mmCIF if they're
present in the PDB chemical component dictionary. Reason for that is
the connectivity information we get from there. Actively enabling this
flag in the compare-ligand-structures action allows to use a distance
based heuristic as fallback if the ligand is not in the dicitonary.