Read SEQRES/SEQRES mapping from mmCIF files in scoring actions
The logic becomes slightly convoluted in this case. Especially with
respect to the residue-number-alignments flag. If the reference
structure is given as mmCIF or if SEQRES is explicitely provided,
the chem grouping is purely based on SEQRES and the respective
mapping. The residue-number-alignments flag has no influence.
If the reference structure is given as PDB, the behavior stays the
same => clustering by sequence identity with pairwise alignments that
are either built using residue numbers or Needleman-Wunsch
(residue-number-alignments flag).