clear separation of concerns in model_check.cc - change in behavior
Several of the checking functions checked for unknown residues and
then set a flag that this check has already been done. When actually
calling CheckForUnknownAtoms this could be skipped. Thats overly
complicated and has been simplified.

One change in behavior is the identification of nonstandard residues.
Previous behavior: If a residue is unknown, the check for nonstandard
was not even performed.
New behavior: A residue can be unknown AND non standard. This is relevant
for molck where I can enable the removal of non-standard residues. I would
expect that a residue with name "im_such_a_standard_guy" is
removed in that case.

However, the molck output now reports this residue as both:
unknown and nonstandard.