MMCIFPrep refactor: Fallbacks if entity/seqres info not in mmCIF file
If only the polymer part is required and no entity info is available
in mmCIF file, a warning is logged and the polymer part falls back to
a selection for protein/nucleotide properties as defined in the
chemical component dictionary
(prot.Select("peptide=true or nucleotide=true"))

If the non-polymers are required and no entity info is available in
mmCIF file, an error is raised. This is only relevant for the
compare-ligand-structures action.

If SEQRES info should be extracted for chem grouping and the mmCIF
file lacks the respective entity_poly_seq categories, a warning is
logged and None returned => In case of the scoring actions, this
triggers the sequence identity based grouping as it was done before.